pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-terminal-cli"
continversion = "0.1.5"
description = "Terminal assistant built as an open source minimal alternative to Warp"
authors = [{name = "Abhinav", email = "abhinavkumarsingh2023@gmail.com"}]
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Shells",
    "Topic :: Utilities",
]
keywords = ["terminal", "ai", "shell", "assistant", "cli", "gemini"]
dependencies = [
    "prompt-toolkit==3.0.51",
    "rich==14.1.0", 
    "pydantic==2.11.7",
    "python-dotenv==1.1.1",
    "google-genai==1.31.0"
]

[project.scripts]
ai-terminal = "terminal.cli:main"

[project.urls]
Homepage = "https://github.com/Abhinavexists/ai-terminal"
Repository = "https://github.com/Abhinavexists/ai-terminal.git"
Issues = "https://github.com/Abhinavexists/ai-terminal/issues"

[tool.setuptools.packages.find]
include = ["terminal*"]
exclude = ["release*", "scripts*", "ai-terminal-linux*"]

[tool.setuptools.package-data]
terminal = ["*.py"]